+++ built/grammar/GrammarExampleNode.d.ts
@@ -1,0 +1,4 @@
+import TreeNode from "../base/TreeNode";
+declare class GrammarExampleNode extends TreeNode {
+}
+export default GrammarExampleNode;