+++ built/grammar/GrammarExampleNode.js
@@ -1,0 +1,6 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const TreeNode_1 = require("../base/TreeNode");
+class GrammarExampleNode extends TreeNode_1.default {
+}
+exports.default = GrammarExampleNode;