+++ built/jtree.js
@@ -19,5 +19,5 @@ jtree.AnyNode = GrammarBackedAnyNode_1.default;
 jtree.GrammarProgram = GrammarProgram_1.default;
 jtree.TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode_1.default;
 jtree.getLanguage = name => require(__dirname + `/../langs/${name}/index.js`);
-jtree.getVersion = () => "19.4.0";
+jtree.getVersion = () => "19.5.0";
 exports.default = jtree;