+++ src/grammar/GrammarExampleNode.ts
@@ -1,0 +1,5 @@
+import TreeNode from "../base/TreeNode"
+
+class GrammarExampleNode extends TreeNode {}
+
+export default GrammarExampleNode